Created simple level selection screen
+ Updated main menu to go to level selection screen
+ Added level select scene to Build settings list
