+++ Catpocalypse/UserSettings/EditorUserSettings.asset
@@ -6,22 +6,25 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
+      value: 5454505453045a0e5456557640770e444e1619292a2970617c7b4b6be3e1656c
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 0609005552570d08080f097516720744124f4c282e7124647a2c1f37b6e1366b
+      value: 5454505453045a0e5456557640770e444e1619292a2970617c7b4b6be3e1656c
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 5454505453045a0e5456557640770e444e1619292a2970617c7b4b6be3e1656c
-      flags: 0
-    RecentlyUsedSceneGuid-3:
-      value: 5454505453045a0e5456557640770e444e1619292a2970617c7b4b6be3e1656c
-      flags: 0
-    RecentlyUsedSceneGuid-4:
       value: 500957045d56085d55565c2713720c444e4f417a7d2b72347f2f4860b2e3313a
       flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-3:
       value: 5a08505f54500c5f5f0a0f7713745a44414e48792a2b24697b7a1e66bae2663c
+      flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
+      flags: 0
+    RecentlyUsedSceneGuid-5:
+      value: 0609005552570d08080f097516720744124f4c282e7124647a2c1f37b6e1366b
+      flags: 0
+    RecentlyUsedSceneGuid-6:
+      value: 075357055605590e5a575c7616270944464f497b2a717e642c78456bb0b76d6c
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
